platform: support non-pci platform devices
This makes things not completely fail if DDX implements platformProbe()
but the device is not actually a PCI device.  Also, the platform device
name does not always match the DDX name, so deal with that.  I'm sure
there are more cases that find_non_pci_driver() needs to handle for
that, but this is a start.

Signed-off-by: Rob Clark <robdclark@gmail.com>
Reviewed-by: Hans de Goede <hdegoede@redhat.com>