add SERVER_SUPPORTS_NON_PCI_PLATFORM_DEVS
Give the DDX a way to know whether non-pci platform devices are
completely broken or not.  For xserver prior to the fix, the
DDX should not claim a platform device in platformProbe(), as
the server will fallback to old ->Probe(), which will fail if
the device is already claimed.  Meaning that a user could not
make things work even with a .conf file to explicitly specify
the driver to use.

Idea suggested by Hans de Goede.

Signed-off-by: Rob Clark <robdclark@gmail.com>
Reviewed-by: Hans de Goede <hdegoede@redhat.com>